replace siae with siad configuration settings
